[SPARK-LLAP-156] Support LlapRelation table size property
This estimated table size would be used for optimizing the join plan. If the estimated size is not big, the broadcast join will be used. Thus, this PR is used to get the estimated table size from metadata s.t. computing optimized joined plan.

```
[root@ctr-e134-1499953498516-99573-01-000002 python]# ./spark-ranger-secure-test.py TableTestSuite
.....................
----------------------------------------------------------------------
Ran 21 tests in 3011.014s

[root@ctr-e134-1499953498516-99573-01-000002 python]# ./spark-ranger-secure-test.py DbTestSuite
....
----------------------------------------------------------------------
Ran 4 tests in 612.001s
```

Author: Dong Joon Hyun <dhyun@hortonworks.com>, Mingjie Tang <mtang@hortonworks.com>